Pre-Dijkstra commit

+++ src/MultiModalRouter/DrivingQuery.java
@@ -1,21 +1,21 @@
 package src.MultiModalRouter;
 
 public class DrivingQuery {
-    private int originNodeId;
-    private int destinationNodeId;
+    private long originNodeId;
+    private long destinationNodeId;
     private int desiredDepartureTimeMinutes;
 
-    DrivingQuery(int originNodeId, int destinationNodeId, int desiredDepartureTimeMinutes) {
+    DrivingQuery(long originNodeId, long destinationNodeId, int desiredDepartureTimeMinutes) {
         this.originNodeId = originNodeId;
         this.destinationNodeId = destinationNodeId;
         this.desiredDepartureTimeMinutes = desiredDepartureTimeMinutes;
     }
 
-    public int getOriginNodeId() {
+    public long getOriginNodeId() {
         return originNodeId;
     }
 
-    public int getDestinationNodeId() {
+    public long getDestinationNodeId() {
         return destinationNodeId;
     }
